Day02/requirements.txt
flake8==7.0.0
psycopg2-binary==2.9.9
numpy==1.26.4
matplotlib==3.9.0
scikit-learn==1.5.0